Swap float as default fft type.
In float-fft, calculate sqrt() on the result
when getting the absolute. Only +10% execution time
(it still is faster than the integer version :))

+++ src/datatype.h
@@ -15,6 +15,7 @@ constexpr unsigned frequency_buffer_samples = listen_buffer_samples / 2;
 typedef std::array<audio_sample_t, listen_buffer_samples> listen_buf_t;
 
 typedef uint16_t fft_bin_t;
-typedef int32_t fft_internal_datatype;
+//typedef int32_t fft_internal_datatype;
+typedef float fft_internal_datatype;
 typedef std::array<fft_bin_t, frequency_buffer_samples> frequency_buf_t;
